Day 32 - Flash Card App Capstone/requirements.txt
pandas==2.3.3
pillow==12.0.0